refactor(position): total rank and file size constants to Position
- fix broken imports
