+++ src/main/java/chess2/domain2/game2/GameResult.java
@@ -4,14 +4,14 @@ import static chess2.domain2.board2.piece2.Color.BLACK;
 import static chess2.domain2.board2.piece2.Color.WHITE;
 import static chess2.domain2.board2.piece2.PieceType.KING;
 import static chess2.domain2.board2.piece2.PieceType.PAWN;
-import static chess2.util2.PositionConverterUtil.FILES_TOTAL_SIZE;
-import static chess2.util2.PositionConverterUtil.RANKS_TOTAL_SIZE;
+import static chess2.domain2.board2.position.Position.FILES_TOTAL_SIZE;
+import static chess2.domain2.board2.position.Position.RANKS_TOTAL_SIZE;
 
-import chess2.domain2.board2.position.Position;
 import chess2.domain2.board2.piece2.Color;
 import chess2.domain2.board2.piece2.Piece;
-import chess2.util2.PieceScoreUtil;
+import chess2.domain2.board2.position.Position;
 import chess2.dto2.GameScoreDto;
+import chess2.util2.PieceScoreUtil;
 import java.util.Map;
 import java.util.stream.IntStream;
